requirements.txt
flask>=2.0.0
click>=8.0.0
orjson>=3.8.0
python-dotenv>=0.19.0
tqdm>=4.65.0
flask-oauthlib>=0.9.6
pathlib>=1.0.1
regex>=2023.0.0